about Page digerible

+++ database/migrations/2021_12_16_142517_create_podio_table.php
@@ -18,7 +18,7 @@ class CreatePodioTable extends Migration
             $table->increments('id');
             $table->unsignedBigInteger('user_id')->nullable();
             $table->string('username');
-            $table->integer('resultado');
+            $table->float('resultado');
             //$table->timestamps();
             $table->timestamp('created_at')->default(DB::raw('CURRENT_TIMESTAMP'));
             $table->timestamp('updated_at')->default(DB::raw('CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP'));
